refactor: move dataTransferObjects out to lib from lib/server/
