+++ src/lib/server/database/storefronts.ts
@@ -1,5 +1,5 @@
-import type { coordinates } from "../dataTransferObjects";
-import { Vendor, Storefront, type MenuItem } from '$lib/server/dataTransferObjects';
+import type { coordinates } from "../../dataTransferObjects";
+import { Vendor, Storefront, type MenuItem } from '$lib/dataTransferObjects';
 import { error, type NumericRange } from '@sveltejs/kit';
 import type { SupabaseClient } from "@supabase/supabase-js";
